To do: - Make the ball bounce off the bar - Implement initGameVars

+++ src/main/java/facade/HomeworkTwoFacade.java
@@ -33,7 +33,7 @@ public class HomeworkTwoFacade {
      * @return a new level determined by the parameters
      * @see Level
      */
-    public Level newLevelWithBricksFull(String name, int numberOfBricks, double probOfGlass, double probOfMetal, int seed) {
+    public Level newLevelWithBricksFull(String name, int numberOfBricks, double probOfGlass, double probOfMetal, long seed) {
         return new PlayableLevel(name, numberOfBricks, probOfGlass, probOfMetal, seed);
     }
 
@@ -47,7 +47,7 @@ public class HomeworkTwoFacade {
      * @return a new level determined by the parameters
      * @see Level
      */
-    public Level newLevelWithBricksNoMetal(String name, int numberOfBricks, double probOfGlass, int seed) {
+    public Level newLevelWithBricksNoMetal(String name, int numberOfBricks, double probOfGlass, long seed) {
         return newLevelWithBricksFull(name, numberOfBricks, probOfGlass, 0.0, seed);
     }
